+++ Sources/KeyboardKitSwiftUI/Extensions/Image+Keyboard.swift
@@ -14,5 +14,4 @@ public extension Image {
     static var backspace: Image { Image(systemName: "delete.left") }
     static var globe: Image { Image(systemName: "globe") }
     static var newLine: Image { Image(systemName: "arrow.turn.down.left") }
-    static var space: Image { Image(systemName: "arrow.right") }
 }